# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Item Test API",
        "description": "API untuk mengelola data dokter phc",
        "version": "1.0"
    },
    "paths": {
...
        "/api/pasien": {
            "get": {
                "tags": [
                    "Pasien"
                ],
                "summary": "Ambil semua data pasien",
                "description": "Mengambil daftar semua pasien yang terdaftar. Bisa menggunakan parameter pencarian berdasarkan ID atau teks.",
                "operationId": "49dd4f9c6782abbdbf414e9da709f276",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Cari berdasarkan ID pasien yang spesifik (integer match).",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Cari berdasarkan NIK, FullName, Phone, atau Address dengan pencarian LIKE.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar pasien berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Patient"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pasien"
                ],
                "summary": "Tambah data pasien baru",
                "description": "Menambahkan pasien baru ke dalam database.",
                "operationId": "933dddcf1a6e640f74ad4a2541f1a008",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NIK",
                                    "PatientID_Provider"
                                ],
                                "properties": {
                                    "NIK": {
                                        "type": "string",
                                        "example": "1234567890123456"
                                    },
                                    "PatientID_Provider": {
                                        "type": "string",
                                        "example": "P-001"
                                    },
                                    "FullName": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "Sex": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    },
                                    "BirthDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "Address": {
                                        "type": "string",
                                        "example": "Jl. Contoh No.1"
                                    },
                                    "Phone": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "CreateBy": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pasien berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Patient created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Patient"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/pasien/{id}": {
            "put": {
                "tags": [
                    "Pasien"
                ],
                "summary": "Perbarui data pasien yang ada",
                "description": "Memperbarui informasi pasien berdasarkan ID yang diberikan.",
                "operationId": "ff1aa41bfd8f5c0034823c6acd18040b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pasien yang akan diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "FullName",
                                    "Sex",
                                    "BirthDate",
                                    "Address",
                                    "Phone",
                                    "LastModifiedBy"
                                ],
                                "properties": {
                                    "FullName": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "Sex": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "F"
                                    },
                                    "BirthDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-05-15"
                                    },
                                    "Address": {
                                        "type": "string",
                                        "example": "Jl. Contoh No.2"
                                    },
                                    "Phone": {
                                        "type": "string",
                                        "example": "081298765432"
                                    },
                                    "LastModifiedBy": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pasien berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Patient updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Patient"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pasien tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Patient": {
                "required": [
                    "NIK",
                    "PatientID_Provider",
                    "FullName",
                    "Sex",
                    "BirthDate"
                ],
                "properties": {
                    "NIK": {
                        "type": "string",
                        "example": "1234567890123456"
                    },
                    "PatientID_Provider": {
                        "type": "string",
                        "example": "P-00123"
                    },
                    "FullName": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "Sex": {
                        "type": "string",
                        "example": "Male"
                    },
                    "BirthDate": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "Address": {
                        "type": "string",
                        "example": "Jl. Sudirman No. 10"
                    },
                    "Phone": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "CreateBy": {
                        "type": "string",
                        "example": "admin"
                    },
                    "LastModifiedBy": {
                        "type": "string",
                        "example": "editor"
                    },
                    "gcrecord": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        }
    }
}